Common interface

+++ src/Disks/HDFS/DiskHDFS.cpp
@@ -27,74 +27,30 @@ namespace ErrorCodes
 {
     extern const int FILE_ALREADY_EXISTS;
     extern const int BAD_ARGUMENTS;
+    extern const int CANNOT_DELETE_DIRECTORY;
+    extern const int UNKNOWN_FORMAT;
+    extern const int PATH_ACCESS_DENIED;
 }
 
 
 DiskHDFS::DiskHDFS(
-    const String & name_,
-    const String & hdfs_name_,
+    const String & disk_name_,
+    const String & hdfs_root_path_,
     const String & metadata_path_,
     ContextPtr context_)
-    : WithContext(context_)
+    : IDiskRemote(hdfs_root_path_, metadata_path_)
     , log(&Poco::Logger::get("DiskHDFS"))
-    , name(name_)
-    , hdfs_name(hdfs_name_)
-    , metadata_path(std::move(metadata_path_))
+    , disk_name(disk_name_)
     , config(context_->getGlobalContext()->getConfigRef())
-    , builder(createHDFSBuilder(hdfs_name, config))
-    , fs(createHDFSFS(builder.get()))
+    , hdfs_builder(createHDFSBuilder(hdfs_root_path_, config))
+    , hdfs_fs(createHDFSFS(hdfs_builder.get()))
 {
-}
-
-
-bool DiskHDFS::exists(const String & path) const
-{
-    return Poco::File(metadata_path + path).exists();
-}
-
-
-bool DiskHDFS::isFile(const String & path) const
-{
-    return Poco::File(metadata_path + path).isFile();
-}
-
-
-bool DiskHDFS::isDirectory(const String & path) const
-{
-    return Poco::File(metadata_path + path).isDirectory();
-}
-
-
-size_t DiskHDFS::getFileSize(const String & path) const
-{
-    Metadata metadata(metadata_path, path);
-    return metadata.total_size;
-}
-
-
-void DiskHDFS::createDirectory(const String & path)
-{
-    Poco::File(metadata_path + path).createDirectory();
-}
-
-
-void DiskHDFS::createDirectories(const String & path)
-{
-    Poco::File(metadata_path + path).createDirectories();
 }
 
 
 DiskDirectoryIteratorPtr DiskHDFS::iterateDirectory(const String & path)
 {
     return std::make_unique<DiskHDFSDirectoryIterator>(metadata_path + path, path);
-}
-
-
-void DiskHDFS::clearDirectory(const String & path)
-{
-    for (auto it{iterateDirectory(path)}; it->isValid(); it->next())
-        if (isFile(it->path()))
-            removeFile(it->path());
 }
 
 
@@ -107,63 +63,53 @@ void DiskHDFS::moveFile(const String & from_path, const String & to_path)
 }
 
 
-void DiskHDFS::replaceFile(const String & from_path, const String & to_path)
-{
-    Poco::File from_file(metadata_path + from_path);
-    Poco::File to_file(metadata_path + to_path);
-    if (to_file.exists())
-    {
-        Poco::File tmp_file(metadata_path + to_path + ".old");
-        to_file.renameTo(tmp_file.path());
-        from_file.renameTo(metadata_path + to_path);
-        removeFile(to_path + ".old");
-    }
-    else
-        from_file.renameTo(to_file.path());
-}
-
-
 std::unique_ptr<ReadBufferFromFileBase> DiskHDFS::readFile(const String & path, size_t buf_size, size_t, size_t, size_t, MMappedFileCache *) const
 {
-    Metadata metadata(metadata_path, path);
+    auto metadata = readMeta(path);
 
     LOG_DEBUG(log,
         "Read from file by path: {}. Existing HDFS objects: {}",
-        backQuote(metadata_path + path), metadata.hdfs_objects.size());
+        backQuote(metadata_path + path), metadata.remote_fs_objects.size());
 
-    return std::make_unique<ReadIndirectBufferFromHDFS>(getContext(), hdfs_name, "", metadata, buf_size);
+    return std::make_unique<ReadIndirectBufferFromHDFS>(config, remote_fs_root_path, "", metadata, buf_size);
 }
 
 
 std::unique_ptr<WriteBufferFromFileBase> DiskHDFS::writeFile(const String & path, size_t buf_size, WriteMode mode)
 {
     bool exist = exists(path);
+
+    if (exist && readMeta(path).read_only)
+        throw Exception(ErrorCodes::PATH_ACCESS_DENIED, "File is read-only: " + path);
+
     /// Path to store new HDFS object.
     auto file_name = getRandomName();
-    auto hdfs_path = hdfs_name + file_name;
+    auto hdfs_path = remote_fs_root_path + file_name;
+
     if (!exist || mode == WriteMode::Rewrite)
     {
         /// If metadata file exists - remove and new.
         if (exist)
             removeFile(path);
-        Metadata metadata(metadata_path, path, true);
+
+        auto metadata = createMeta(path);
         /// Save empty metadata to disk to have ability to get file size while buffer is not finalized.
         metadata.save();
 
         LOG_DEBUG(log,
             "Write to file by path: {}. New hdfs path: {}", backQuote(metadata_path + path), hdfs_path);
 
-        return std::make_unique<WriteIndirectBufferFromHDFS>(getContext(), hdfs_path, file_name, metadata, buf_size);
+        return std::make_unique<WriteIndirectBufferFromHDFS>(config, hdfs_path, file_name, metadata, buf_size);
     }
     else
     {
-        Metadata metadata(metadata_path, path);
+        auto metadata = readMeta(path);
 
         LOG_DEBUG(log,
             "Append to file by path: {}. New hdfs path: {}. Existing HDFS objects: {}",
-            backQuote(metadata_path + path), hdfs_path, metadata.hdfs_objects.size());
+            backQuote(metadata_path + path), hdfs_path, metadata.remote_fs_objects.size());
 
-        return std::make_unique<WriteIndirectBufferFromHDFS>(getContext(), hdfs_path, file_name, metadata, buf_size);
+        return std::make_unique<WriteIndirectBufferFromHDFS>(config, hdfs_path, file_name, metadata, buf_size);
     }
 }
 
@@ -173,21 +119,27 @@ void DiskHDFS::removeFile(const String & path)
     LOG_DEBUG(&Poco::Logger::get("DiskHDFS"), "Remove file by path: {}", backQuote(metadata_path + path));
 
     Poco::File file(metadata_path + path);
-    if (file.isFile())
+
+    if (!file.isFile())
+        throw Exception(ErrorCodes::CANNOT_DELETE_DIRECTORY, "Path '{}' is a directory", path);
+
+    try
     {
-        Metadata metadata(metadata_path, path);
+        auto metadata = readMeta(path);
 
         /// If there is no references - delete content from HDFS.
         if (metadata.ref_count == 0)
         {
             file.remove();
-            for (const auto & [hdfs_object_path, _] : metadata.hdfs_objects)
+
+            for (const auto & [hdfs_object_path, _] : metadata.remote_fs_objects)
             {
-                const size_t begin_of_path = hdfs_name.find('/', hdfs_name.find("//") + 2);
-                const std::string hdfs_path = hdfs_name.substr(begin_of_path) + hdfs_object_path;
-                int res = hdfsDelete(fs.get(), hdfs_path.c_str(), 0);
+                const size_t begin_of_path = remote_fs_root_path.find('/', remote_fs_root_path.find("//") + 2);
+                const String hdfs_path = remote_fs_root_path.substr(begin_of_path) + hdfs_object_path;
+
+                int res = hdfsDelete(hdfs_fs.get(), hdfs_path.c_str(), 0);
                 if (res == -1)
-                    throw Exception("HDFSDelete failed with path: " + hdfs_path, 1);
+                    throw Exception(ErrorCodes::LOGICAL_ERROR, "HDFSDelete failed with path: " + hdfs_path);
             }
         }
         else /// In other case decrement number of references, save metadata and delete file.
@@ -197,16 +149,32 @@ void DiskHDFS::removeFile(const String & path)
             file.remove();
         }
     }
-    else
-        file.remove();
+    catch (const Exception & e)
+    {
+        /// If it's impossible to read meta - just remove it from FS.
+        if (e.code() == ErrorCodes::UNKNOWN_FORMAT)
+        {
+            LOG_WARNING(
+                log,
+                "Metadata file {} can't be read by reason: {}. Removing it forcibly.",
+                backQuote(path),
+                e.nested() ? e.nested()->message() : e.message());
+
+            file.remove();
+        }
+        else
+            throw;
+    }
 }
 
 
 void DiskHDFS::removeFileIfExists(const String & path)
 {
-    int exists_status = hdfsExists(fs.get(), path.data());
+    int exists_status = hdfsExists(hdfs_fs.get(), path.data());
     if (exists_status == 0)
         removeFile(path);
+    else
+        LOG_DEBUG(&Poco::Logger::get("DiskHDFS"), "File by path {} does not exist", backQuote(metadata_path + path));
 }
 
 
@@ -228,54 +196,15 @@ void DiskHDFS::removeRecursive(const String & path)
 }
 
 
-void DiskHDFS::removeDirectory(const String & path)
-{
-    Poco::File(metadata_path + path).remove();
-}
-
-
-void DiskHDFS::listFiles(const String & path, std::vector<String> & file_names)
-{
-    for (auto it = iterateDirectory(path); it->isValid(); it->next())
-        file_names.push_back(it->name());
-}
-
-
-void DiskHDFS::setLastModified(const String & path, const Poco::Timestamp & timestamp)
-{
-    Poco::File(metadata_path + path).setLastModified(timestamp);
-}
-
-
-Poco::Timestamp DiskHDFS::getLastModified(const String & path)
-{
-    return Poco::File(metadata_path + path).getLastModified();
-}
-
-
 void DiskHDFS::createHardLink(const String & src_path, const String & dst_path)
 {
     /// Increment number of references.
-    Metadata src(metadata_path, src_path);
+    auto src = readMeta(src_path);
     ++src.ref_count;
     src.save();
 
     /// Create FS hardlink to metadata file.
     DB::createHardLink(metadata_path + src_path, metadata_path + dst_path);
-}
-
-
-void DiskHDFS::createFile(const String & path)
-{
-    /// Create empty metadata file.
-    Metadata metadata(metadata_path, path, true);
-    metadata.save();
-}
-
-
-void DiskHDFS::setReadOnly(const String & path)
-{
-    Poco::File(metadata_path + path).setReadOnly(true);
 }
 
 
@@ -293,7 +222,7 @@ bool DiskHDFS::tryReserve(UInt64 bytes)
 
     if (bytes == 0)
     {
-        LOG_DEBUG(log, "Reserving 0 bytes on HDFS disk {}", backQuote(name));
+        LOG_DEBUG(log, "Reserving 0 bytes on HDFS disk {}", backQuote(disk_name));
         ++reservation_count;
         return true;
     }
@@ -304,7 +233,7 @@ bool DiskHDFS::tryReserve(UInt64 bytes)
     {
         LOG_DEBUG(log,
             "Reserving {} on disk {}, having unreserved {}",
-            formatReadableSizeWithBinarySuffix(bytes), backQuote(name), formatReadableSizeWithBinarySuffix(unreserved_space));
+            formatReadableSizeWithBinarySuffix(bytes), backQuote(disk_name), formatReadableSizeWithBinarySuffix(unreserved_space));
         ++reservation_count;
         reserved_bytes += bytes;
         return true;
@@ -319,49 +248,23 @@ void registerDiskHDFS(DiskFactory & factory)
     auto creator = [](const String & name,
                       const Poco::Util::AbstractConfiguration & config,
                       const String & config_prefix,
-                      ContextConstPtr context) -> DiskPtr
+                      ContextConstPtr context_) -> DiskPtr
     {
-        Poco::File disk{context->getPath() + "disks/" + name};
+        Poco::File disk{context_->getPath() + "disks/" + name};
         disk.createDirectories();
 
-        DB::String uri{config.getString(config_prefix + ".endpoint")};
+        String uri{config.getString(config_prefix + ".endpoint")};
 
         if (uri.back() != '/')
             throw Exception(ErrorCodes::BAD_ARGUMENTS, "HDFS path must ends with '/', but '{}' doesn't.", uri);
 
-        String metadata_path = context->getPath() + "disks/" + name + "/";
-        auto copy_context = Context::createCopy(context);
-        return std::make_shared<DiskHDFS>(name, uri, metadata_path, copy_context);
+        String metadata_path = context_->getPath() + "disks/" + name + "/";
+        auto context = Context::createCopy(context_);
+
+        return std::make_shared<DiskHDFS>(name, uri, metadata_path, context);
     };
 
     factory.registerDiskType("hdfs", creator);
 }
-
-
-//void DiskHDFS::copyFile(const String & from_path, const String & to_path)
-//{
-//    if (exists(to_path))
-//        remove(to_path);
-//
-//    Metadata from(metadata_path, from_path);
-//    Metadata to(metadata_path, to_path, true);
-//
-//    for (const auto & [path, size] : from.hdfs_objects)
-//    {
-//        auto new_path = hdfs_name + getRandomName();
-//        /// TODO:: hdfs copy semantics
-//        /*
-//        Aws::HDFS::Model::CopyObjectRequest req;
-//        req.SetCopySource(bucket + "/" + path);
-//        req.SetBucket(bucket);
-//        req.SetKey(new_path);
-//        throwIfError(client->CopyObject(req));
-//        */
-//        throw Exception("is not implemented yet", 1);
-//        to.addObject(new_path, size);
-//    }
-//
-//    to.save();
-//}
 
 }
